Pulled out itemIter into own file, split into two structs
