+++ store/btree.go
@@ -170,19 +170,7 @@ func (b BTreeCacheWrap) Iterator(start, end []byte) (Iterator, error) {
 	if err != nil {
 		return nil, err
 	}
-	iter := newItemIter(parentIter)
-
-	if start == nil && end == nil {
-		b.bt.Ascend(iter.insert)
-	} else if start == nil { // end != nil
-		b.bt.AscendLessThan(bkey{end}, iter.insert)
-	} else if end == nil { // start != nil
-		b.bt.AscendGreaterOrEqual(bkey{start}, iter.insert)
-	} else { // both != nil
-		b.bt.AscendRange(bkey{start}, bkey{end}, iter.insert)
-	}
-	iter.skipAllDeleted()
-
+	iter := ascendBtree(b.bt, start, end).wrap(parentIter)
 	return iter, nil
 }
 
@@ -194,19 +182,7 @@ func (b BTreeCacheWrap) ReverseIterator(start, end []byte) (Iterator, error) {
 	if err != nil {
 		return nil, err
 	}
-	iter := newItemIter(parentIter)
-
-	if start == nil && end == nil {
-		b.bt.Descend(iter.insert)
-	} else if start == nil { // end != nil
-		b.bt.DescendLessOrEqual(bkeyLess{end}, iter.insert)
-	} else if end == nil { // start != nil
-		b.bt.DescendGreaterThan(bkeyLess{start}, iter.insert)
-	} else { // both != nil
-		b.bt.DescendRange(bkeyLess{end}, bkeyLess{start}, iter.insert)
-	}
-	iter.skipAllDeleted()
-
+	iter := descendBtree(b.bt, start, end).wrap(parentIter)
 	return iter, nil
 }
 
@@ -276,184 +252,4 @@ func (k bkeyLess) Key() []byte {
 func (k bkeyLess) Less(item btree.Item) bool {
 	cmp := item.(keyer).Key()
 	return bytes.Compare(k.key, cmp) <= 0
-}
-
-///////////////////////////////////////////////////////
-// From Items to Iterator
-
-// TODO: add support for Combine (deleting those below)
-type itemIter struct {
-	data []btree.Item
-	idx  int
-	// if we are iterating in a cache-wrap (and who isn't),
-	// we need to combine this iterator with the parent
-	parent Iterator
-}
-
-var _ Iterator = (*itemIter)(nil)
-
-// source marks where the current item comes from
-type source int32
-
-const (
-	us source = iota
-	parent
-	both
-	none
-)
-
-// combine joins our results with those of the parent,
-// taking into consideration overwrites and deletes...
-func newItemIter(parent Iterator) *itemIter {
-	return &itemIter{
-		parent: parent,
-	}
-}
-
-// insert is designed as a callback to add items from the btree.
-// Example Usage (to get an iterator over all items on the tree):
-//
-//  iter := newItemIter(parentIter)
-//  tree.Ascend(iter.insert)
-//  iter.skipAllDeleted()
-func (i *itemIter) insert(item btree.Item) bool {
-	i.data = append(i.data, item)
-	return true
-}
-
-// makes sure the parent is non-nil before checking if it is valid
-func (i *itemIter) parentValid() bool {
-	return (i.parent != nil) && i.parent.Valid()
-}
-
-// makes sure the parent is non-nil before checking if it is valid
-func (i *itemIter) weValid() bool {
-	return (i.idx < len(i.data))
-}
-
-//------- public facing interface ------
-
-// Valid implements Iterator and returns true iff it can be read
-func (i *itemIter) Valid() bool {
-	return i.weValid() || i.parentValid()
-}
-
-// Next moves the iterator to the next sequential key in the database, as
-// defined by order of iteration.
-//
-// If Valid returns false, this method will panic.
-func (i *itemIter) Next() error {
-	// advance either us, parent, or both
-	switch i.firstKey() {
-	case us:
-		i.idx++
-	case both:
-		i.idx++
-		fallthrough
-	case parent:
-		err := i.parent.Next()
-		if err != nil {
-			return err
-		}
-	default:
-		panic("Advanced past the end!")
-	}
-
-	// keep advancing over all deleted entries
-	return i.skipAllDeleted()
-}
-
-// Key returns the key of the cursor.
-func (i *itemIter) Key() (key []byte) {
-	switch i.firstKey() {
-	case us, both:
-		return i.get().Key()
-	case parent:
-		return i.parent.Key()
-	default: //none
-		panic("Advanced past the end!")
-	}
-}
-
-// Value returns the value of the cursor.
-func (i *itemIter) Value() (value []byte) {
-	switch i.firstKey() {
-	case us, both:
-		return i.get().(setItem).value
-	case parent:
-		return i.parent.Value()
-	default: // none
-		panic("Advanced past the end!")
-	}
-}
-
-// Close releases the Iterator.
-func (i *itemIter) Close() {
-	i.data = nil
-}
-
-// skipAllDeleted loops and skips any number of deleted items
-func (i *itemIter) skipAllDeleted() error {
-	var err error
-	more := true
-	for more {
-		more, err = i.skipDeleted()
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// skipDeleted jumps over all elements we can safely fast forward
-// return true if skipped, so we can skip again
-func (i *itemIter) skipDeleted() (bool, error) {
-	src := i.firstKey()
-	if src == us || src == both {
-		// if our next is deleted, advance...
-		if _, ok := i.get().(deletedItem); ok {
-			i.idx++
-			// if parent had the same key, advance parent as well
-			if src == both {
-				err := i.parent.Next()
-				if err != nil {
-					return false, err
-				}
-			}
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-// get requires this is valid, gets what we are pointing at
-func (i *itemIter) get() keyer {
-	return i.data[i.idx].(keyer)
-}
-
-// firstKey selects the iterator with the lowest key is any
-func (i *itemIter) firstKey() source {
-	// if only one or none is valid, it is clear which to use
-	if !i.parentValid() {
-		if !i.weValid() {
-			return none
-		}
-		return us
-	} else if !i.weValid() {
-		return parent
-	}
-
-	// both are valid... compare keys....
-	parKey := i.parent.Key()
-	usKey := i.get().Key()
-
-	// let's see which one to do....
-	cmp := bytes.Compare(parKey, usKey)
-	if cmp < 0 {
-		return parent
-	} else if cmp > 0 {
-		return us
-	} else {
-		return both
-	}
 }
